some updates for the PR
